Included details about this only working with Teensy 2.0
Thanks for the great guide.  I tried this with a Teensy 3.1 and confirmed that it does not work.  Rather than trying to write a new ISP programming package for the Teensy 3.1, I bought a Teensy 2.0 and successfully flashed bootloaders on two Tokyo60 v3 PCBs.

Please have a look at this proposed change to make it clear that later versions of the Teensy should not be bought for the purpose of fixing bootloaders.